services/payments/pyproject.toml
[project]
name = "eci-payments"
version = "1.0.0"
description = "payments microservice"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.0",
  "sqlalchemy==2.0.35",
  "psycopg2-binary==2.9.9",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "alembic==1.13.2",
  "redis==5.0.1",
  "psutil==5.9.8",
  "python-json-logger==2.0.7",
  "httpx==0.26.0",
  "pytest==8.0.0",
  "pytest-asyncio==0.23.3",
  "pytest-cov==4.1.0"
]
